fix: the position widening was incomplete, and add the missing regression tests
Two float locals caught a double return and re-narrowed it, so the position
fix in 3fa91c08 and 81454eb8 did not actually reach the reported value:

  PottsScorer.cpp  const float direct  = score_direct(...)
  PWMScorer.cpp    float result        = try_slide_window(...)

Both sit on the path a fresh interval takes, so every reported position went
back through float regardless of the widened signatures. PWMScorer's is in
code already merged to master (#177), so that fix shipped incomplete.

I did not catch it because both verifications used an EVEN target position.
binary32 represents even integers exactly up to 2^25, so 16777218 comes back
right whether or not the fix works; only an ODD position past 2^24 exercises
it. The new test plants its anchor at an odd offset and now reports 16777525
where the old code reported 16777524.

Five regression tests, one per fix that had none - the `$` partial match, the
flat-J shape check at npair == 0, the tightened float bound, the vtrack
position past 2^24, and the iterator error message. Each verified to FAIL
against a build without its fix and pass with it.

The position test needs an interval longer than 16.7 Mb and so a chromosome
longer than that; the fixture's largest is 500 kb and materialising a 17 Mb one
would cost every CI job on every platform for a two-base error. It skips unless
MISHA_LARGE_GENOME points at a real assembly. It also searches for an anchor
that occurs ONCE - a first version asserted only that the maximum was reached
and passed while reporting position 3042193, because the planted 20-mer sat in
a repeat and the scan found an earlier copy.

An earlier draft of the same test scored a long bare string through
gseq.potts(). That was powerless too: C_gseq_potts builds its position as
`(double)(best_i + 1)` and never reached the defect, so the test passed against
the unfixed build.

+++ src/PWMScorer.cpp
@@ -816,7 +816,7 @@ double PWMScorer::score_with_sliding_window(const std::string& target,
 
     // Try to slide if possible
     if (can_slide) {
-        float result = try_slide_window(target, original_interval, expanded_interval, i_min, i_max, motif_len, stride);
+        double result = try_slide_window(target, original_interval, expanded_interval, i_min, i_max, motif_len, stride);
         if (!std::isnan(result)) {
             return result;
         }
